Test séparation .cpp et .h pour gestionaireTimer2
