+++ rover_arduino/platformio.ini
@@ -13,4 +13,7 @@ platform = atmelavr
 board = megaatmega2560
 framework = arduino
 monitor_speed = 57600
-
+lib_deps=
+    Servo
+    MsTimer2
+    ros_lib_melodic
