+++ rover_arduino/src/gestionaireTimer2.cpp
@@ -1,24 +1,4 @@
-#ifndef GESTIONAIRETIMER2_H
-#define GESTIONAIRETIMER2_H
-
-//#include <MsTimer2.h>
-
-
-class gestionaireTimer2
-{
-private:
-    static void f0();
-    static void f1();
-    static void f2();
-    static void f3();
-    static void f4();
-    static void(*fonctions[5])();
-    static int taille ;
-public:
-    gestionaireTimer2();
-    static void lancer();
-    static void ajouter(void(*g)());
-};
+#include "gestionaireTimer2.h"
 
 int gestionaireTimer2::taille = 0;
 void(*gestionaireTimer2::fonctions[5])() {gestionaireTimer2::f0,gestionaireTimer2::f1,gestionaireTimer2::f2,gestionaireTimer2::f3,gestionaireTimer2::f4};
@@ -49,6 +29,3 @@ void gestionaireTimer2::ajouter(void(*nouvelleFonction)())
     MsTimer2::set(5,lancer);
     MsTimer2::start();
 }
-
-
-#endif
